vim-patch:8.2.4241: some type casts are redundant
Problem:    Some type casts are redundant.
Solution:   Remove the type casts. (closes vim/vim#9643)
https://github.com/vim/vim/commit/420fabcd4ffeaf79082a6e43db91e1d363f88f27

This is not a literal port but an equivalent one.

+++ src/nvim/debugger.c
@@ -268,7 +268,7 @@ void do_debug(char_u *cmd)
                        DOCMD_VERBOSE|DOCMD_EXCRESET);
       debug_break_level = n;
     }
-    lines_left = (int)(Rows - 1);
+    lines_left = Rows - 1;
   }
   xfree(cmdline);
 
@@ -277,7 +277,7 @@ void do_debug(char_u *cmd)
   redraw_all_later(NOT_VALID);
   need_wait_return = false;
   msg_scroll = save_msg_scroll;
-  lines_left = (int)(Rows - 1);
+  lines_left = Rows - 1;
   State = save_State;
   debug_mode = false;
   did_emsg = save_did_emsg;
